Merged PR 87719: Rebind static dates
## Description
![image.png](https://dev.azure.com/jibe-erp/c85e9e04-673e-4f17-8fce-3dab78037534/_apis/git/repositories/52124beb-a908-434d-bbba-8a3454873a88/pullRequests/87719/attachments/image.png)
General idea is that dates/time shall appear the same regardless the timezones.

For example:

set the date of the PC to GMT+1
edit a statement of facts and put in date and time (eg 20231201 10:30)
change TZ to India TZ
go to same sow, it shall show 20231201 10:30

## Testing

<!--- Please give a description of any manual testing steps that might be required. -->

## Breaking changes

<!--- Does this PR cause breaking changes that might impact other developers, please describe them here and what other
developers might have to accommodate for. -->

## Checklist
<!---
Please remove any checklist items not relevant to your PR
-->

- [ ] [Synchronization calls](https://dev.azure.com/jibe-erp/JiBe/_workitems/edit/685837) are added (if you added any code to modify the table)
- [ ] This PR fulfills all acceptance criteria
- [ ] Relevant well written tests included

Related work items: #702943