Merged PR 87710: Fixed the way we send errors to frontend and how we handle them on frontend
## Description

Previously it was a little bit a mess and improper way to handle the errors. We treat all errors except 422 as internal ones. Also, we were checking wrong fields on the frontend to find the explanation of the error.

That's fixed now. Now the backend always replies with "name" and/or "message" of the error. That's generic properties of the error that most (? at least ApiRequestService of jibe-components) JiBe code expects. Then on frontend we handle this error via generic handler and user sees a nice explanation of the error, not only validation/server error messages.

@<Ivan Chenskyi UA Dev> @<Mikhail.Aulov>  FYI

Related work items: #658790