Merged PR 87671: Project Monitoring - Gantt Chart (PHASE I)
## Description

Refers to the addition of a non-interactive Gantt chart in the area named as "Gantt Chart".

The Grid area of the chart will be displaying the following columns:

Task Name - it will display the specification subject
Start Date - it will display the specification start date
End Date - it will display the specification end date
Status - (this will be the workflow status of the specification)
Progress - it will display the percentage completion as defined in the job orders
The Gantt on the top will display the month and the relevant days (as per Figma)
The Gantt chart will consider the "Start Date" and the "End Date" so as to form the relevant graph.
The colors on the rows of the chart will be identical to the colors of the status.## Testing

<!--- Please give a description of any manual testing steps that might be required. -->

## Breaking changes

<!--- Does this PR cause breaking changes that might impact other developers, please describe them here and what other
developers might have to accommodate for. -->

## Checklist
<!---
Please remove any checklist items not relevant to your PR
-->

- [ ] [Synchronization calls](https://dev.azure.com/jibe-erp/JiBe/_workitems/edit/685837) are added (if you added any code to modify the table)
- [ ] This PR fulfills all acceptance criteria
- [ ] Relevant well written tests included

Related work items: #673281